+++ app/controllers/client_api/users/users_controller.rb
@@ -1,0 +1,16 @@
+module ClientApi
+  module Users
+    class UsersController < ApplicationController
+
+      def current_user_info
+        respond_to do |format|
+          format.json do
+            render template: '/client_api/users/show',
+                  status: :ok,
+                  locals: { user: current_user }
+          end
+        end
+      end
+    end
+  end
+end
